requirements.txt
flask
stripe
python-dotenv
flask-cors
gunicorn